+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.3.2] - 2021-11-02
+
+- Added
+  - Support for sharing DA app bundle or activity aliases with other Forge apps
+  - Retrieving details of a specific DA alias (for both app bundles and activities)
+
 ## [2.3.1] - 2021-06-03
 
 - Added